+++ spring-cloud-huawei-router-client/src/main/java/com/huaweicloud/router/client/loabalancer/RouterLoadBalancerClientConfiguration.java
@@ -59,7 +59,7 @@ import reactor.util.retry.RetrySpec;
  */
 @Configuration(proxyBeanMethods = false)
 @ConditionalOnDiscoveryEnabled
-public class CanaryLoadBalancerClientConfiguration {
+public class RouterLoadBalancerClientConfiguration {
 
   private static final int REACTIVE_SERVICE_INSTANCE_SUPPLIER_ORDER = 193827465;
 
@@ -85,7 +85,7 @@ public class CanaryLoadBalancerClientConfiguration {
         matchIfMissing = true)
     public ServiceInstanceListSupplier discoveryClientServiceInstanceListSupplier(
         ConfigurableApplicationContext context) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
           ServiceInstanceListSupplier.builder().withDiscoveryClient().withCaching().build(context));
     }
 
@@ -95,7 +95,7 @@ public class CanaryLoadBalancerClientConfiguration {
     @ConditionalOnProperty(value = "spring.cloud.loadbalancer.configurations", havingValue = "zone-preference")
     public ServiceInstanceListSupplier zonePreferenceDiscoveryClientServiceInstanceListSupplier(
         ConfigurableApplicationContext context) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
           ServiceInstanceListSupplier.builder().withDiscoveryClient().withZonePreference().withCaching()
               .build(context));
     }
@@ -106,7 +106,7 @@ public class CanaryLoadBalancerClientConfiguration {
     @ConditionalOnProperty(value = "spring.cloud.loadbalancer.configurations", havingValue = "health-check")
     public ServiceInstanceListSupplier healthCheckDiscoveryClientServiceInstanceListSupplier(
         ConfigurableApplicationContext context) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
           ServiceInstanceListSupplier.builder().withDiscoveryClient().withHealthChecks().build(context));
     }
 
@@ -117,7 +117,7 @@ public class CanaryLoadBalancerClientConfiguration {
         havingValue = "request-based-sticky-session")
     public ServiceInstanceListSupplier requestBasedStickySessionDiscoveryClientServiceInstanceListSupplier(
         ConfigurableApplicationContext context) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
           ServiceInstanceListSupplier.builder().withDiscoveryClient().withRequestBasedStickySession()
               .build(context));
     }
@@ -129,7 +129,7 @@ public class CanaryLoadBalancerClientConfiguration {
         havingValue = "same-instance-preference")
     public ServiceInstanceListSupplier sameInstancePreferenceServiceInstanceListSupplier(
         ConfigurableApplicationContext context) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
           ServiceInstanceListSupplier.builder().withDiscoveryClient().withSameInstancePreference()
               .build(context));
     }
@@ -148,7 +148,7 @@ public class CanaryLoadBalancerClientConfiguration {
             matchIfMissing = true)
     public ServiceInstanceListSupplier discoveryClientServiceInstanceListSupplier(
             ConfigurableApplicationContext context, ServiceCombRegistration serviceCombRegistration) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
               new ZoneAwareServiceInstanceListSupplier(
                       ServiceInstanceListSupplier.builder().withDiscoveryClient().withCaching().build(context),
                       serviceCombRegistration));
@@ -160,7 +160,7 @@ public class CanaryLoadBalancerClientConfiguration {
     @ConditionalOnProperty(value = "spring.cloud.loadbalancer.configurations", havingValue = "zone-preference")
     public ServiceInstanceListSupplier zonePreferenceDiscoveryClientServiceInstanceListSupplier(
             ConfigurableApplicationContext context, ServiceCombRegistration  serviceCombRegistration) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
               new ZoneAwareServiceInstanceListSupplier(
                       ServiceInstanceListSupplier.builder().withDiscoveryClient().withZonePreference().withCaching()
                               .build(context), serviceCombRegistration));
@@ -172,7 +172,7 @@ public class CanaryLoadBalancerClientConfiguration {
     @ConditionalOnProperty(value = "spring.cloud.loadbalancer.configurations", havingValue = "health-check")
     public ServiceInstanceListSupplier healthCheckDiscoveryClientServiceInstanceListSupplier(
             ConfigurableApplicationContext context, ServiceCombRegistration  serviceCombRegistration) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
               new ZoneAwareServiceInstanceListSupplier(
                       ServiceInstanceListSupplier.builder().withDiscoveryClient().withHealthChecks().build(context),
                       serviceCombRegistration));
@@ -185,7 +185,7 @@ public class CanaryLoadBalancerClientConfiguration {
             havingValue = "request-based-sticky-session")
     public ServiceInstanceListSupplier requestBasedStickySessionDiscoveryClientServiceInstanceListSupplier(
             ConfigurableApplicationContext context, ServiceCombRegistration  serviceCombRegistration) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
               new ZoneAwareServiceInstanceListSupplier(
                       ServiceInstanceListSupplier.builder().withDiscoveryClient().withRequestBasedStickySession()
                               .build(context), serviceCombRegistration));
@@ -198,7 +198,7 @@ public class CanaryLoadBalancerClientConfiguration {
             havingValue = "same-instance-preference")
     public ServiceInstanceListSupplier sameInstancePreferenceServiceInstanceListSupplier(
             ConfigurableApplicationContext context, ServiceCombRegistration  serviceCombRegistration) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
               new ZoneAwareServiceInstanceListSupplier(
                       ServiceInstanceListSupplier.builder().withDiscoveryClient().withSameInstancePreference()
                               .build(context), serviceCombRegistration));
@@ -218,7 +218,7 @@ public class CanaryLoadBalancerClientConfiguration {
         matchIfMissing = true)
     public ServiceInstanceListSupplier discoveryClientServiceInstanceListSupplier(
         ConfigurableApplicationContext context) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
           ServiceInstanceListSupplier.builder().withBlockingDiscoveryClient().withCaching().build(context));
     }
 
@@ -228,7 +228,7 @@ public class CanaryLoadBalancerClientConfiguration {
     @ConditionalOnProperty(value = "spring.cloud.loadbalancer.configurations", havingValue = "zone-preference")
     public ServiceInstanceListSupplier zonePreferenceDiscoveryClientServiceInstanceListSupplier(
         ConfigurableApplicationContext context) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
           ServiceInstanceListSupplier.builder().withBlockingDiscoveryClient().withZonePreference()
               .withCaching().build(context));
     }
@@ -239,7 +239,7 @@ public class CanaryLoadBalancerClientConfiguration {
     @ConditionalOnProperty(value = "spring.cloud.loadbalancer.configurations", havingValue = "health-check")
     public ServiceInstanceListSupplier healthCheckDiscoveryClientServiceInstanceListSupplier(
         ConfigurableApplicationContext context) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
           ServiceInstanceListSupplier.builder().withBlockingDiscoveryClient().withBlockingHealthChecks()
               .build(context));
     }
@@ -251,7 +251,7 @@ public class CanaryLoadBalancerClientConfiguration {
         havingValue = "request-based-sticky-session")
     public ServiceInstanceListSupplier requestBasedStickySessionDiscoveryClientServiceInstanceListSupplier(
         ConfigurableApplicationContext context) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
           ServiceInstanceListSupplier.builder().withBlockingDiscoveryClient().withRequestBasedStickySession()
               .build(context));
     }
@@ -263,7 +263,7 @@ public class CanaryLoadBalancerClientConfiguration {
         havingValue = "same-instance-preference")
     public ServiceInstanceListSupplier sameInstancePreferenceServiceInstanceListSupplier(
         ConfigurableApplicationContext context) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
           ServiceInstanceListSupplier.builder().withBlockingDiscoveryClient().withSameInstancePreference()
               .build(context));
     }
@@ -282,7 +282,7 @@ public class CanaryLoadBalancerClientConfiguration {
             matchIfMissing = true)
     public ServiceInstanceListSupplier discoveryClientServiceInstanceListSupplier(
             ConfigurableApplicationContext context, ServiceCombRegistration  serviceCombRegistration) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
               new ZoneAwareServiceInstanceListSupplier(
                       ServiceInstanceListSupplier.builder().withBlockingDiscoveryClient().withCaching().build(context),
                       serviceCombRegistration));
@@ -294,7 +294,7 @@ public class CanaryLoadBalancerClientConfiguration {
     @ConditionalOnProperty(value = "spring.cloud.loadbalancer.configurations", havingValue = "zone-preference")
     public ServiceInstanceListSupplier zonePreferenceDiscoveryClientServiceInstanceListSupplier(
             ConfigurableApplicationContext context, ServiceCombRegistration  serviceCombRegistration) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
               new ZoneAwareServiceInstanceListSupplier(
                       ServiceInstanceListSupplier.builder().withBlockingDiscoveryClient().withZonePreference()
                               .withCaching().build(context), serviceCombRegistration));
@@ -306,7 +306,7 @@ public class CanaryLoadBalancerClientConfiguration {
     @ConditionalOnProperty(value = "spring.cloud.loadbalancer.configurations", havingValue = "health-check")
     public ServiceInstanceListSupplier healthCheckDiscoveryClientServiceInstanceListSupplier(
             ConfigurableApplicationContext context, ServiceCombRegistration  serviceCombRegistration) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
               new ZoneAwareServiceInstanceListSupplier(
                       ServiceInstanceListSupplier.builder().withBlockingDiscoveryClient().withBlockingHealthChecks()
                               .build(context), serviceCombRegistration));
@@ -319,7 +319,7 @@ public class CanaryLoadBalancerClientConfiguration {
             havingValue = "request-based-sticky-session")
     public ServiceInstanceListSupplier requestBasedStickySessionDiscoveryClientServiceInstanceListSupplier(
             ConfigurableApplicationContext context, ServiceCombRegistration  serviceCombRegistration) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
               new ZoneAwareServiceInstanceListSupplier(
                       ServiceInstanceListSupplier.builder().withBlockingDiscoveryClient().withRequestBasedStickySession()
                               .build(context), serviceCombRegistration));
@@ -332,7 +332,7 @@ public class CanaryLoadBalancerClientConfiguration {
             havingValue = "same-instance-preference")
     public ServiceInstanceListSupplier sameInstancePreferenceServiceInstanceListSupplier(
             ConfigurableApplicationContext context, ServiceCombRegistration  serviceCombRegistration) {
-      return new CanaryServiceInstanceListSupplier(
+      return new RouterServiceInstanceListSupplier(
               new ZoneAwareServiceInstanceListSupplier(
                       ServiceInstanceListSupplier.builder().withBlockingDiscoveryClient().withSameInstancePreference()
                               .build(context), serviceCombRegistration));
@@ -342,7 +342,7 @@ public class CanaryLoadBalancerClientConfiguration {
   @Configuration(proxyBeanMethods = false)
   @ConditionalOnBlockingDiscoveryEnabled
   @ConditionalOnClass(RetryTemplate.class)
-  @Conditional(CanaryLoadBalancerClientConfiguration.BlockingOnAvoidPreviousInstanceAndRetryEnabledCondition.class)
+  @Conditional(RouterLoadBalancerClientConfiguration.BlockingOnAvoidPreviousInstanceAndRetryEnabledCondition.class)
   @AutoConfigureAfter(LoadBalancerClientConfiguration.BlockingSupportConfiguration.class)
   @ConditionalOnBean(ServiceInstanceListSupplier.class)
   public static class BlockingRetryConfiguration {
@@ -358,7 +358,7 @@ public class CanaryLoadBalancerClientConfiguration {
 
   @Configuration(proxyBeanMethods = false)
   @ConditionalOnBlockingDiscoveryEnabled
-  @Conditional(CanaryLoadBalancerClientConfiguration.ReactiveOnAvoidPreviousInstanceAndRetryEnabledCondition.class)
+  @Conditional(RouterLoadBalancerClientConfiguration.ReactiveOnAvoidPreviousInstanceAndRetryEnabledCondition.class)
   @AutoConfigureAfter(LoadBalancerClientConfiguration.ReactiveSupportConfiguration.class)
   @ConditionalOnBean(ServiceInstanceListSupplier.class)
   @ConditionalOnClass(RetrySpec.class)
